Initial commit of FleetCommander plugin for Observatory Core
This simple plugin's primary purpose is to tell you when your carrier's jump completes and when the jump cooldown ends with verbal + visual notifications (since the game can't be bothered to tell you).

Note that, currently Odyssey is not reporting "CarrierJump" events in the journal so jumps which complete while you are not docked will not be reported and thus also can't be tracked for cooldown. However, this will work quite fine when docked and doing many carrier jumps back-to-back.

Enjoy!