Fix number parsing errors in locales which use ',' for decimal separator
The galactic records is written using US formatting, notably, using a period for decimal separator, not a comma. Attempting to import these values using a locale which uses a comma as decimal separator (ie. de or ru) causes lots of errors to be written to the crash log on startup.

This change forces US locale when parsing the galactic records file.