[Stat Scanner] Fix for RegionsVisitedRecord and add belts EDAstro table
RegionsVisitedRecord tracks the set of regions you've visited. However, it was implemented under the assumption that the data given to it in the constructor is also the data from the database -- which it is not. As such, the "ExtraData" value from the database was being ignored and ultimately overridden on subsequent program runs resulting in mis-fires and inconsistency between the value and ExtraData. This is now read from the inner data object when checking the codex entry every time rather only on record object construction.

Also, the latest version of the galactic records files have records from the "belts" table. Adding it to the RecordsTable enum to avoid debug noise. We currently don't use/track these records.