[Aggregator] New plugin work-in-progress.
NOTE: Full functionality of this plugin depends on unreleased features in Xjph/ObservatoryCore. Therefore, there is no release.

This plugin collects notifications from other plugins and shows them all in 1 view to reduce the amount of flipping between plugins that is necessary.

This is a fairly functional starting point, but I have lots more ideas to implement once Core provides the functionality.