[FleetCommander] Jump Cooldown timer works when not docked
I discovered that the CarrierJumpRequest provides a time for the scheduled jump, so I can determine the cooldown from that vs. a Jump event in the journal (which isn't provided if you're not docked). Furthermore, a restart of ObservatoryCore should pick up the in-progress countdown (if it's recent enough) and resume correctly.

In general, the cooldown should work far more reliably now.

This plus a few other misc bug fixes.