[AutoUpdater] Add Core Min Required version support
Automatic updates can now be blocked based on the version of Core being used. This will be needed for many plugins once the next Core update lands as there will be new features in the framework which won't be available in older versions. Closed a gap in logic that could prevent an update to the latest production version when beta versions are also enabled. Plus some further refactoring and tests.

Also made the CurrentVersion.json parsing more robust by allowing trailing commas (which I have a knack for including via copy & paste).