[FleetCommander] Estimate fuel used where possible, add some workarounds
The game does not provide either jump distance nor fuel usage for carrier jumps. Where there is accurate location data (and in realtime mode, it will try to fetch coordinates from edastro), it will compute fuel requirements for the jump and update the fuel level to avoid excess or delayed Fuel Level change update (often triggered by opening Carrier Management when plotting the next jump). This leverages a global cache to avoid frequent requests and to backfill for low-fidelity gaps later if revisiting a system with the carrier. Fuel usage is not computed when doing an in-system jump because often this change is reverted by a subsequent CarrierStats event...

Also added logic to work around the missing CarrierJump events, etc. by using unresolved CarrierJumpRequests as the trigger for the carrier jump complete event. However, these are unreliable and do not raise notifications.