[StatScanner] UI Overhaul support: notification coalescing + bugfixes
Add support for the new notification coalescing feature in the upcoming version of Core. The idea is that some notification listeners (like Aggregator) can group related things from different plugins together.

To avoid spamming notifications early in a commander's career, bodies which generate a lot of new personal bests (ie. the first of an object ever found), these notifications are squashed down into a single "multiple personal bests" notification now. The # of notifications to trigger squashing is a new setting.

Fixed a nagging bug with the Undiscovered System Tally which caused it to not properly reset if you did multiple read-alls in a single session.

Also apply default column widths (new in the upcoming ObsCore UI overhaul as well).