[Aggregator] Significant updates, UI Overhaul support, new features
Aggregator now does grouping/coalescing based on IDs sent by other plugins. Displays Sender id when provided. For each body with a notification, it also creates a summary row of the body (optionally doing so for every body, regardless of if there's a notification for it). There is also a system summary row. This entailed a non-trivial amount of refactoring.

To accomplish the above, aggregator now re-draws the grid content on every change which makes it less than ideal to show anything but the current system (that setting is removed). That said, Read-All data is now buffered entirely until complete (but this fact is only really useful in Dev mode where notifications can be sent.

Notifications from plugins like Explorer which are combined are split into separate lines (because ListView doesn't support multiple lines).

Now listens for data from Archivist, but doesn't yet do anything with it. Need to come up with a clean way of de-serializing.

Bonus: Updates to the TestCore impl to reflect changes to the api.