[Archivist] Implement experimental data sharing w/ setting
Xjph/ObservatoryCore#137 added grouping , double up/down and other settings layout improvements. Use that with a new setting for enabling data sharing via the new message bus in Core. This is still experimental; not sold on the data structure I'm using yet -- it may be overly complicated to parse. This is likely to change.