[AutoUpdater] Upgrade to .net8 and implement a UI
An overhaul which features a full UI that shows you the current state of installed plugins, allows you to manually trigger an update or install plugins that are not currently installed. The UI also provides messaging about activities and outcomes. Also stashes some state locally to reduce round-trips to GH (and to seed the UI with non-empty data).

Plus upgrade to .net8.