[AutoUpdater] UI fixes for next release of Core
A recent change broke the UI (it was creating many MANY instances of the UI panel vs. a single instance, and eventually was just... blank). Fixed that, but also made some changes to improve resize layout performance a bit. I'm still not fully happy with it. The TableLayoutPanel can be really sluggish sometimes.

Also implement the new ObservatoryReady method to trigger checking for updates.