[StatScanner] Use the new LogMonitorState.BatchCancelled plus more
The new state allows the plugin to capture when a read-all is cancelled and persist the read-all requirement/reason.

Also used the new ObservatoryReady handler to show the initial content (vs. watching for pre-read finishing).

Furthermore, the summary now includes some stats for all known commanders, not just the current commander, but only displays the current commander summary of records (although this will change eventually too).