[AutoUpdater] DataGridView implementation; fixes bugs
Discovered you would be offered a button to install, but if you didn't have betas enabled and there was no produciton release, the button did nothing and didn't say anything.

Rewrote the UI with a DataGridView and it's soo much better. The Use Beta setting is now on the plugin UI and dynamically updates the grid content if toggled (and keeps in sync with the settings page).

It works with older versions of ObsCore, but may not look pretty with some themes until ObsCore >= 0.4 releases.