[FleetCommander] Fix UI-blocking issue due to Spansh response changes
This update features appropriate use of concurrent programming facilities to avoid blocking the entire app UI while waiting for a result from spansh....  Now, if Spansh takes a while, you can actually go grab the JobID out of the text box OR cancel (imagine that)...

Also expanded the error message a bit to indicate unknown systems as a cause of route plotting failures. Oh, and trim input to avoid pointless failures. Y'know, like we should.

Thanks @mcmuttons for reporting this!