[FleetCommander] Fix error handling, support multiple stops, try import
Addresses an issue reported on discord. Easily reproducible by providing a garbage destination system. Also removes the error logging for this to avoid pop-ups given that I have textbox for error output already.

Allow route plotting to accept multiple destinations and handle the duplicate entry in the list of jumps for each waypoint.

Also, add support for importing a route from a spansh url. However, leave it disabled for now (button is not visible). The window of opportunity for this is ~15 min.