Chore: Fix versioning to handle the turn-of-the-year better
With the old versioning scheme, newly compiled versions are all less than the latest release on github -- meaning auto-updater kept pulling the github version down, thinking they were newer. This is because the revision value rolled back to a smaller number than previous because it's based on day of year + hour. So, force increment the build number to 1.0.4 AND update the logic for the Revision number to be yy + day of year -- which never rolls back to a lower value at the turn of the year. Downside: It only generates 1 version/day. (Kinda annoying that the maximum value for each part of the version is 65536.)

Also, simplify the version specification a whole lot. And mark Archivist and Helm as beta.

Anyway, I'm good for another 40 years of revisions... That's more than enough, right?